Refs #24682: Storyblok - move base and plugin interfaces into the main module
