--- components/storyblok/interfaces/alignment.value.ts
@@ -1,5 +1,0 @@
-export enum Alignment {
-  LEFT = 'left',
-  CENTER = 'center',
-  RIGHT = 'right'
-}
