--- components/storyblok/interfaces/asset-field.interface.ts
@@ -1,10 +1,0 @@
-export default interface AssetField {
-  alt: string,
-  copyright: string,
-  fieldtype: string,
-  filename: string,
-  focus?: string,
-  id: number,
-  name: string,
-  title: string
-}
