+++ components/storyblok/interfaces/button-item-data.interface.ts
@@ -1,7 +1,6 @@
-import Item from './item.interface';
-import LinkField from './link-field.interface';
+import { ItemData, LinkField } from 'src/modules/vsf-storyblok-module';
 
-export default interface ButtonItemData extends Item {
+export default interface ButtonItemData extends ItemData {
   link_text: string,
   link_url: LinkField,
   is_primary: boolean,
