+++ components/storyblok/interfaces/buttons-data.interface.ts
@@ -1,6 +1,6 @@
 import ButtonItemData from './button-item-data.interface';
-import Item from './item.interface';
+import { ItemData } from 'src/modules/vsf-storyblok-module';
 
-export default interface ButtonsData extends Item {
+export default interface ButtonsData extends ItemData {
   button_items: ButtonItemData[]
 }
