+++ components/storyblok/interfaces/category-data.interface.ts
@@ -1,6 +1,6 @@
-import Item from './item.interface';
+import { ItemData } from 'src/modules/vsf-storyblok-module';
 
-export default interface CategoryData extends Item {
+export default interface CategoryData extends ItemData {
   id: string,
   products_count: string
 }
