--- components/storyblok/interfaces/color-picker-field.interface.ts
@@ -1,3 +1,0 @@
-export default interface ColorPickerField {
-  color: string
-}
