--- components/storyblok/interfaces/display.value.ts
@@ -1,4 +1,0 @@
-export enum Display {
-  MOBILE_HIDDEN = 'mobile_hidden',
-  MOBILE_ONLY = 'mobile_only',
-}
