+++ components/storyblok/interfaces/heading-data.interface.ts
@@ -1,6 +1,6 @@
-import Item from './item.interface';
+import { ItemData } from 'src/modules/vsf-storyblok-module';
 
-export default interface HeadingData extends Item {
+export default interface HeadingData extends ItemData {
   title: string,
   heading_type: string
 }
