--- components/storyblok/interfaces/link-type.value.ts
@@ -1,4 +1,0 @@
-export enum LinkType {
-  STORY = 'story',
-  URL = 'url',
-}
