+++ components/storyblok/interfaces/promo-campaign-banner-data.interface.ts
@@ -1,5 +1,5 @@
-import Item from './item.interface';
+import { ItemData } from 'src/modules/vsf-storyblok-module';
 
-export default interface PromoCampaignBannerData extends Item {
+export default interface PromoCampaignBannerData extends ItemData {
   campaign_id: string
 }
