+++ components/storyblok/interfaces/sharing-data.interface.ts
@@ -1,8 +1,6 @@
-import Item from './item.interface';
-import AssetField from './asset-field.interface';
-import LinkField from './link-field.interface';
+import { AssetField, ItemData, LinkField } from 'src/modules/vsf-storyblok-module';
 
-export default interface SharingData extends Item {
+export default interface SharingData extends ItemData {
   sharing_url: LinkField,
   sharing_description: string,
   twitter_description: string,
