--- components/storyblok/interfaces/spacing-settings-field.interface.ts
@@ -1,10 +1,0 @@
-interface SpacingSettingsField {
-  margin_top: string,
-  margin_right: string,
-  margin_bottom: string,
-  margin_left: string,
-  padding_top: string,
-  padding_right: string,
-  padding_bottom: string,
-  padding_left: string
-}
