+++ components/storyblok/interfaces/textarea-data.interface.ts
@@ -1,6 +1,6 @@
-import Item from './item.interface';
+import { ItemData } from 'src/modules/vsf-storyblok-module';
 
-export default interface TextareaData extends Item {
+export default interface TextareaData extends ItemData {
   type: string,
   content: unknown[]
 }
