+++ components/storyblok/interfaces/video-data.interface.ts
@@ -1,8 +1,7 @@
 import { AspectRatio } from './aspect-ratio.value';
-import Item from './item.interface';
-import VideoUrlField from './video-url-field.interface';
+import { ItemData, VideoUrlField } from 'src/modules/vsf-storyblok-module';
 
-export default interface VideoData extends Item {
+export default interface VideoData extends ItemData {
   url: VideoUrlField,
   aspect_ratio: AspectRatio,
   display_controls?: boolean
