--- components/storyblok/interfaces/video-provider.value.ts
@@ -1,12 +1,0 @@
-export enum VideoProvider {
-  youtube = 'youtube',
-  vimeo = 'vimeo',
-  twitch = 'twitch',
-  dailymotion = 'dailymotion',
-  canalplus = 'canalplus',
-  youku = 'youku',
-  coub = 'coub',
-  wistia = 'wistia',
-  soundcloud = 'soundcloud',
-  teachertube = 'teachertube',
-}
