--- components/storyblok/interfaces/video-url-field.interface.ts
@@ -1,7 +1,0 @@
-import { VideoProvider } from './video-provider.value';
-
-export default interface VideoUrlField {
-  video_id: string,
-  video_url: string,
-  provider: VideoProvider
-}
