+++ components/storyblok/interfaces/website-rating-data.interface.ts
@@ -1,7 +1,6 @@
-import Item from './item.interface';
-import LinkField from './link-field.interface';
+import { ItemData, LinkField } from 'src/modules/vsf-storyblok-module';
 
-export default interface WebsiteRatingData extends Item {
+export default interface WebsiteRatingData extends ItemData {
   link_url: LinkField,
   link_text: string
 }
